implement a driver capability - FAST_HIEARACHY
