+++ maestro-client/src/main/java/maestro/Capability.kt
@@ -1,0 +1,5 @@
+package maestro
+
+enum class Capability {
+    FAST_HIERARCHY,
+}
