+++ maestro-client/src/main/java/maestro/drivers/IOSDriver.kt
@@ -28,6 +28,7 @@ import hierarchy.IdbElementNode
 import hierarchy.XCUIElement
 import hierarchy.XCUIElementNode
 import ios.IOSDevice
+import maestro.Capability
 import maestro.DeviceInfo
 import maestro.Driver
 import maestro.KeyCode
@@ -41,7 +42,6 @@ import maestro.ViewHierarchy
 import maestro.utils.ScreenshotUtils
 import maestro.utils.FileUtils
 import maestro.utils.MaestroTimer
-import maestro.utils.TapUtils
 import okio.Sink
 import org.slf4j.LoggerFactory
 import util.XCRunnerSimctl
@@ -457,30 +457,9 @@ class IOSDriver(
         return if (didFinishOnTime) null else ScreenshotUtils.waitForAppToSettle(initialHierarchy, this)
     }
 
-//    override fun performTap(x: Int,
-//                            y: Int,
-//                            retryIfNoChange: Boolean,
-//                            longPress: Boolean,
-//                            initialHierarchy: ViewHierarchy?) {
-//        LOGGER.info("Tapping at ($x, $y)")
-//
-//        val hierarchyBeforeTap = initialHierarchy ?: TapUtils.viewHierarchy(this)
-//
-//        val retries = TapUtils.getNumberOfRetries(retryIfNoChange)
-//        repeat(retries) {
-//            if (longPress) {
-//                longPress(Point(x, y))
-//            } else {
-//                tap(Point(x, y))
-//            }
-//            val hierarchyAfterTap = waitForAppToSettle(initialHierarchy)
-//
-//            if (hierarchyAfterTap == null || hierarchyBeforeTap != hierarchyAfterTap) {
-//                LOGGER.info("Something have changed in the UI judging by view hierarchy. Proceed.")
-//                return
-//            }
-//        }
-//    }
+    override fun capabilities(): List<Capability> {
+        return emptyList()
+    }
 
     private fun isScreenStatic(): Boolean {
         return iosDevice.isScreenStatic().expect {}
